Add new vars. Implement init.

+++ Headers/AppKit/NSMovieView.h
@@ -44,9 +44,9 @@ APPKIT_EXPORT_CLASS
 @interface NSMovieView : NSView
 {
   @protected
-    NSMovie* _movie;
-    float _rate;
-    float _volume;
+    NSMovie *_movie;
+    CGFloat  _rate;
+    CGFloat  _volume;
     struct NSMovieViewFlags {
       unsigned int muted: 1;
       unsigned int loopMode: 3;
